+++ libs/iso-bmff/src/IsoBoxWriteView.ts
@@ -1,4 +1,5 @@
 import { encodeText } from '@svta/cml-utils'
+import { TEMPLATE } from './fields/TEMPLATE.ts'
 import { UINT } from './fields/UINT.ts'
 import type { IsoFieldTypeMap } from './IsoFieldTypeMap.ts'
 
@@ -186,11 +187,13 @@ export class IsoBoxWriteView {
 	 * @param data - The data to write.
 	 * @param type - The type of the data.
 	 * @param size - The size, in bytes, of each data value.
+	 * @param length - The number of values to write. (optional)
 	 */
-	writeArray = <T extends keyof IsoFieldTypeMap>(data: number[], type: T, size: number): void => {
-		const write = type === UINT ? this.writeUint : this.writeInt
-		for (const value of data) {
-			write(value, size)
+	writeArray = <T extends keyof IsoFieldTypeMap>(data: number[], type: T, size: number, length: number): void => {
+		const write = type === UINT ? this.writeUint : type === TEMPLATE ? this.writeTemplate : this.writeInt
+
+		for (let i = 0; i < length; i++) {
+			write(data[i] ?? 0, size)
 		}
 	}